1787bug修改

+++ app/src/main/java/qx/app/freight/qxappfreight/bean/WeightWayBillBean.java
@@ -61,7 +61,7 @@ public class WeightWayBillBean implements Serializable {
     private String waybillCode;
     private int number;
     private int weight;
-    private int volume;
+    private double volume;
     private String runTimeScooterId;
     private String scooterId;
     private int scooterType;
